requirements.txt
asyncio==3.4.3
blinker==1.7.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.7
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
hiredis==2.2.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
python-dateutil==2.8.2
python-dotenv==1.0.0
redis==5.0.1
requests==2.31.0
schedule==1.2.0
six==1.16.0
SQLAlchemy==2.0.23
typing_extensions==4.8.0
urllib3==2.0.3
websockets==11.0.3
Werkzeug==3.0.1
